refactor: update imports to use agenda-specific types
