+++ pages/agenda.vue
@@ -3,7 +3,8 @@ import CoverTitle from "~/components/cover/CoverTitle.vue"
 import HeroCover from "~/components/cover/HeroCover.vue"
 import { useCalendarEvents } from "~/composables/calendar/use-calendar-events.composable"
 import { EVENT_TYPE_ITEMS } from "~/constants"
-import type { ContentType, EventTypeItem } from "~/types"
+import type { ContentType } from "~/types"
+import type { EventTypeItem } from "~/types/agenda"
 import AgendaFilters from "~/components/agenda/CalendarFilters.vue"
 import CalendarEventList from "~/components/agenda/CalendarEventList.vue"
 import CalendarTrends from "~/components/agenda/CalendarTrends.vue"
